tests/integration/apps/tortoiseorm-app/requirements.txt
starlette
uvicorn
tortoise-orm[asyncpg]
